Refactoring

+++ src/main/java/com/itcube/journal/service/CourseService.java
@@ -1,6 +1,6 @@
 package com.itcube.journal.service;
 
-import com.itcube.journal.domain.Course;
+import com.itcube.journal.model.Course;
 import com.itcube.journal.repos.CourseRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,7 +14,7 @@ public class CourseService {
         return courseRepo.findAll();
     }
 
-    public Course find(Integer id) {
+    public Course find(Long id) {
         return courseRepo.findById(id).get();
     }
 }
